Fix file structure
